requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
SQLAlchemy==2.0.30
pydantic==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.2.0
httpx==0.27.0

